Added a simple shell script to build the kernel
The file kernel-build/custom_kernel_build_vbox.sh is a simple
shell script that automates the process of building a Linux
kernel image from the sources using customized .config file
and it creates an easily installable tarball in the build
folder.

This script is well annotated with detailed comments on each
step of the build process covering some of the best practices
to be followed in building a kernel.